+++ game/game_network.cpp
@@ -925,7 +925,7 @@ void idGameLocal::UpdateLagometer( int aheadOfServer, int dupeUsercmds ) {
 		for ( i = 0; i < LAGO_HEIGHT; i++ ) {
 			lagometer[i][j][0] = lagometer[i][j][1] = lagometer[i][j][2] = lagometer[i][j][3] = 0;
 		}
-		ahead = idMath::Rint( (float)aheadOfServer / 16.0f );
+		ahead = static_cast<int>(idMath::Rint( (float)aheadOfServer / 16.0f ));
 		if ( ahead >= 0 ) {
 			for ( i = 2 * Max( 0, 5 - ahead ); i < 2 * 5; i++ ) {
 				lagometer[i][j][1] = 255;
